Move countdown to separate component
